+++ src/filter/mode3afilter.cpp
@@ -149,9 +149,9 @@ bool Mode3AFilter::activeInLiveMode()
     return true;
 }
 
-std::vector<size_t> Mode3AFilter::filterBuffer(const std::string& dbcontent_name, std::shared_ptr<Buffer> buffer)
+std::vector<unsigned int> Mode3AFilter::filterBuffer(const std::string& dbcontent_name, std::shared_ptr<Buffer> buffer)
 {
-    std::vector<size_t> to_be_removed;
+    std::vector<unsigned int> to_be_removed;
 
     if (!COMPASS::instance().dbContentManager().metaVariable(DBContent::meta_var_m3a_.name()).existsIn(dbcontent_name))
         return to_be_removed;
